# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "AlternativesResponse": {
        "additionalProperties": false,
        "description": "The bounded list of alternative source candidates for a food item.",
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SourceCandidateDTO"
            },
            "title": "Candidates",
            "type": "array"
          }
        },
        "required": [
          "candidates"
        ],
        "title": "AlternativesResponse",
        "type": "object"
      },
...
      "DerivedFoodItemDTO": {
        "description": "Read shape for a food item: current values, original snapshot, provenance.\n\nReturned by the edit ``PATCH`` and every Today read path. ``source`` and\n``is_edited`` are server-derived (FTY-092): ``source`` from the item's\n``evidence_sources`` row (``None`` when no provenance record exists), and\n``is_edited`` ``True`` iff the item carries a ``user_edit`` value-override\ncorrection \u2014 an amount-adjusted-only item stays ``False``.",
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "calories": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories"
          },
          "calories_estimated": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calories Estimated"
          },
          "carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G"
          },
          "carbs_g_estimated": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbs G Estimated"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G"
          },
          "fat_g_estimated": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fat G Estimated"
          },
          "grams": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grams"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_edited": {
            "default": false,
            "title": "Is Edited",
            "type": "boolean"
          },
          "item_type": {
            "const": "food",
            "default": "food",
            "title": "Item Type",
            "type": "string"
          },
          "log_event_id": {
            "format": "uuid",
            "title": "Log Event Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G"
          },
          "protein_g_estimated": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protein G Estimated"
          },
          "quantity_text": {
            "title": "Quantity Text",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemSourceDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/DerivedItemStatus"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "log_event_id",
          "name",
          "quantity_text",
          "unit",
          "amount",
          "status",
          "grams",
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g",
          "calories_estimated",
          "protein_g_estimated",
          "carbs_g_estimated",
          "fat_g_estimated",
          "created_at",
          "updated_at"
        ],
        "title": "DerivedFoodItemDTO",
        "type": "object"
      },
...
      "DerivedItemStatus": {
        "description": "Resolution status of a derived food/exercise item (FTY-042, FTY-196).\n\nA candidate is persisted :attr:`UNRESOLVED` \u2014 parsed from the log text but not\nyet costed. The calculation steps (FTY-043 exercise burn, FTY-044 food\nresolution) later attach calories/macros and advance it to :attr:`RESOLVED`.\nFTY-042 only ever writes :attr:`UNRESOLVED`.\n\n:attr:`PROPOSED` (FTY-196) is a costed-but-**unconfirmed** food item: a legible\nnutrition-label parse holds its computed calories/macros in this state instead\nof :attr:`RESOLVED`, because \"OCR is fallible \u2014 Fatty never silently trusts a\nfallible parse\" (``docs/design-philosophy.md``). A ``proposed`` item is excluded\nfrom every finalized-state read **by construction** (the daily-summary filter\nrequires :attr:`RESOLVED`, see ``docs/contracts/daily-summary.md``), so it never\ncounts toward totals; confirming the proposal transitions it ``proposed \u2192\nresolved`` and it then counts. Only the label path writes :attr:`PROPOSED`; a\ntext parse resolves straight to :attr:`RESOLVED` as before.\n\n``status`` is persisted as a plain ``VARCHAR`` (not a database ``ENUM``), so\nadding :attr:`PROPOSED` is an application-only change with **no** schema\nmigration \u2014 the value round-trips on both SQLite and Postgres unchanged.",
        "enum": [
          "unresolved",
          "resolved",
          "proposed"
        ],
        "title": "DerivedItemStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemSourceDTO": {
        "additionalProperties": false,
        "description": "Per-item provenance descriptor for the Today timeline + sheet (FTY-092).\n\nDerived **read-only** from the item's ``evidence_sources`` row \u2014 there is no\npersisted provenance column. The client renders the always-on source icon and\nlabel from these three fields without joining ``evidence_sources`` /\n``corrections`` itself:\n\n- ``source_type`` \u2014 the evidence-hierarchy classification; a ``model_prior``\n  value is the signal to render the \"\u2248 rough estimate \u00b7 make it exact\" treatment.\n- ``label`` \u2014 a human, display-ready string mapped deterministically from\n  ``source_type`` / ``ref`` (e.g. \"USDA\", \"Open Food Facts\", \"Label scan\", the\n  official-source host, \"Rough estimate\").\n- ``ref`` \u2014 the stable ``source_ref`` (e.g. ``usda_fdc:<id>``,\n  ``open_food_facts:<barcode>``, ``official_source:<url>``) for the sheet's\n  deeper provenance line. For an ``official_source`` item this is the URL only.\n- ``estimate_basis`` \u2014 *(FTY-281, optional)* how a ``user_text`` item's **missing**\n  macros were filled. ``comparable_reference`` marks a **rough comparable-reference\n  aggregate** estimate, so a client can tell it from a plain user-stated item whose\n  macros are unknown; ``None`` (the common case) when no such aggregate backs the\n  item. The item's ``source_type`` stays ``user_text`` \u2014 the calories are still the\n  user's stated number; only the macro estimate carries this secondary basis.",
        "properties": {
          "estimate_basis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MacroEstimateBasis"
              },
              {
                "type": "null"
              }
            ]
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          }
        },
        "required": [
          "source_type",
          "label",
          "ref"
        ],
        "title": "ItemSourceDTO",
        "type": "object"
      },
...
      "ListAlternativesRequest": {
        "additionalProperties": false,
        "description": "Request body for listing a food item's alternative source candidates.\n\n``query`` is an optional item-identity override (the corrected term) so the user\ncan re-aim to a different food, not just re-rank the original name. It is item\nidentity only \u2014 the service sends it through ``sanitize_query`` (control-stripped,\nlength-bounded) before any provider egress; profile, history, and metrics have no\nchannel here. Omitted \u21d2 the providers are queried with the item's own name.",
        "properties": {
          "query": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "title": "ListAlternativesRequest",
        "type": "object"
      },
...
      "MacroEstimateBasis": {
        "description": "How a ``user_text`` item's *missing* macros were filled (FTY-281 read-model).\n\nA user-stated calorie item keeps ``source_type = user_text`` (its calories are the\nnumber the user typed); this secondary signal \u2014 surfaced on\n:class:`~app.schemas.corrections.ItemSourceDTO.estimate_basis` \u2014 lets a client tell\na **rough comparable-reference aggregate** macro estimate from a plain user-stated\nitem whose macros are unknown, so the two never read as the same evidence.\n``None`` on the DTO means no comparable-reference aggregate backs the item.",
        "enum": [
          "comparable_reference"
        ],
        "title": "MacroEstimateBasis",
        "type": "string"
      },
...
      "ReResolveRequest": {
        "additionalProperties": false,
        "description": "Request body for re-resolving a food item to a chosen candidate.\n\n``source_ref`` is the opaque candidate reference the listing step surfaced \u2014 a\n**reference only**, never nutrition values. ``extra = forbid`` rejects any attempt\nto pass facts directly, so a client cannot inject calories/macros through this\npath; the server re-derives the facts from the reference.",
        "properties": {
          "source_ref": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Source Ref",
            "type": "string"
          }
        },
        "required": [
          "source_ref"
        ],
        "title": "ReResolveRequest",
        "type": "object"
      },
...
      "SourceCandidateDTO": {
        "additionalProperties": false,
        "description": "One alternative source match offered for an item (FTY-093).\n\nThe client renders these as \"Change match\" choices: ``source_type`` (the evidence\nclassification driving the source icon), ``source_ref`` (the opaque id echoed back\non re-resolve \u2014 never facts), a display ``name``, the ``basis`` the facts preview\nis expressed against, and the canonical per-basis facts (``calories`` + macros) as\na compact preview. A rough/`partial` or energy-less match is never offered.",
        "properties": {
          "basis": {
            "enum": [
              "per_100g",
              "per_100ml",
              "per_serving"
            ],
            "title": "Basis",
            "type": "string"
          },
          "calories": {
            "title": "Calories",
            "type": "number"
          },
          "carbs_g": {
            "title": "Carbs G",
            "type": "number"
          },
          "fat_g": {
            "title": "Fat G",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protein_g": {
            "title": "Protein G",
            "type": "number"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          }
        },
        "required": [
          "source_type",
          "source_ref",
          "name",
          "basis",
          "calories",
          "protein_g",
          "carbs_g",
          "fat_g"
        ],
        "title": "SourceCandidateDTO",
        "type": "object"
      },
...
      "SourceType": {
        "description": "Evidence source-hierarchy classification for a resolved item (FTY-045/092).\n\nThe canonical vocabulary recorded on each ``evidence_sources`` row's\n``source_type`` (see ``docs/contracts/evidence-retrieval.md``). FTY-092 reads it\ninto the per-item ``source`` descriptor the Today timeline renders, so the client\nmaps a source icon + label without re-deriving the hierarchy. ``model_prior``\nsurfaces a rough estimate plainly so the client can render the\n\"\u2248 rough estimate \u00b7 make it exact\" treatment.",
        "enum": [
          "trusted_nutrition_database",
          "product_database",
          "official_source",
          "user_label",
          "user_text",
          "reference_source",
          "model_prior"
        ],
        "title": "SourceType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "fatty-backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/users/{user_id}/derived-items/food/{item_id}/re-resolve": {
      "post": {
        "description": "Re-resolve the caller's own food item to a chosen candidate reference.\n\nRecomputes calories/macros from the chosen source at the item's current portion,\nrewrites its provenance to the new source, and re-snapshots the estimated\noriginals \u2014 the item is **not** marked edited. The response carries the new\nper-item ``source`` descriptor (FTY-092). Cross-user or unknown items \u2014 and items\nwhose parent log event is voided (FTY-321) \u2014 fail closed as ``404``; a reference\nthe server cannot re-derive, or a re-match the new source cannot cost, returns\n``422`` with a clear error shape.",
        "operationId": "re_resolve_item_api_users__user_id__derived_items_food__item_id__re_resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DerivedFoodItemDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Re Resolve Item",
        "tags": [
          "re-match"
        ]
      }
    },
    "/api/users/{user_id}/derived-items/food/{item_id}/source-candidates": {
      "post": {
        "description": "List alternative source candidates for the caller's own food item.\n\nRuns the existing resolution providers in list-candidates mode over the item's\nidentity (or the sanitized ``query`` override) and returns a bounded list of\nenergy-bearing matches. Cross-user or unknown items \u2014 and items whose parent\nlog event is voided (FTY-321) \u2014 fail closed as ``404``; a transient or unusable\ncandidate-source failure returns ``503`` (retryable) rather than a misleading\nempty list.",
        "operationId": "list_source_candidates_api_users__user_id__derived_items_food__item_id__source_candidates_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListAlternativesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlternativesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Source Candidates",
        "tags": [
          "re-match"
        ]
      }
    },
...
  }
}